Refactor generate plugin data
GeneratePluginData was becoming cumbersome and hard to follow. This
splits parsing into more discrete and logical parts. It hopefully makes
it easier to add more parsing in the future
